+++ src/main/java/com/example/conference/persistence/dao/service/intefaces/TalkService.java
@@ -15,9 +15,9 @@ public interface TalkService {
 
     void saveTalk(Talk Talk);
 
-    Talk getTalk(int id);
+    Talk getTalk(long id);
 
-    void deleteTalk(int id);
+    void deleteTalk(long id);
 
 
 }
